rename vendor --> lib, assuming vendor name is special
