+++ internal/lib/README.md
