+++ internal/lib/darwin-amd64/vendor.go
