+++ internal/lib/darwin-arm64/vendor.go
