+++ internal/lib/doc.go
@@ -3,12 +3,5 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build darwin && arm64 && !go1.22
-package vendor
-
-import _ "embed" // Needed for go:embed
-
-//go:embed darwin-arm64/libddwaf.dylib
-var libddwaf []byte
-
-const embedNamePattern = "libddwaf-*.dylib"
+// Package lib provides a built-in WAF library version for the relevant runtime platform.
+package lib
