+++ internal/lib/lib.go
@@ -5,7 +5,7 @@
 
 //go:build ((darwin && (amd64 || arm64)) || (linux && (amd64 || arm64))) && !go1.22
 
-package vendor
+package lib
 
 import (
 	"fmt"
