+++ internal/lib/lib_darwin_arm64.go
@@ -3,12 +3,12 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build darwin && amd64 && !go1.22
-package vendor
+//go:build darwin && arm64 && !go1.22
+package lib
 
 import _ "embed" // Needed for go:embed
 
-//go:embed darwin-amd64/libddwaf.dylib
+//go:embed darwin-arm64/libddwaf.dylib
 var libddwaf []byte
 
 const embedNamePattern = "libddwaf-*.dylib"
