+++ internal/lib/lib_linux_amd64.go
@@ -3,12 +3,12 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux && arm64 && !go1.22
-package vendor
+//go:build linux && amd64 && !go1.22
+package lib
 
 import _ "embed" // Needed for go:embed
 
-//go:embed linux-arm64/libddwaf.so
+//go:embed linux-amd64/libddwaf.so
 var libddwaf []byte
 
 const embedNamePattern = "libddwaf-*.so"
