+++ internal/lib/lib_linux_arm64.go
@@ -3,5 +3,12 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package vendor provides a built-in WAF library version for the relevant runtime platform.
-package vendor
+//go:build linux && arm64 && !go1.22
+package lib
+
+import _ "embed" // Needed for go:embed
+
+//go:embed linux-arm64/libddwaf.so
+var libddwaf []byte
+
+const embedNamePattern = "libddwaf-*.so"
