+++ internal/lib/linux-amd64/vendor.go
