+++ waf_dl_test.go
@@ -15,7 +15,7 @@ import (
 	"runtime"
 	"testing"
 
-	"github.com/DataDog/go-libddwaf/internal/vendor"
+	"github.com/DataDog/go-libddwaf/internal/lib"
 	"github.com/stretchr/testify/require"
 )
 
@@ -33,7 +33,7 @@ func TestVerifyHeader(t *testing.T) {
 // testVerifyELFHeader is here to ease the debug cases that will likely need
 // to dive in the linker to debug because the error handling is very poor
 func testVerifyELFHeader(t *testing.T) {
-	file, err := vendor.DumpEmbeddedWAF()
+	file, err := lib.DumpEmbeddedWAF()
 	require.NoError(t, err)
 
 	defer func() {
@@ -60,7 +60,7 @@ func testVerifyELFHeader(t *testing.T) {
 // testVerifyMachOHeader is here to ease the debug cases that will likely need
 // to dive in the linker to debug because the error handling is very poor
 func testVerifyMachOHeader(t *testing.T) {
-	file, err := vendor.DumpEmbeddedWAF()
+	file, err := lib.DumpEmbeddedWAF()
 	require.NoError(t, err)
 
 	defer func() {
